backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pypdf==3.17.4
python-docx==0.8.11
langchain==0.0.353
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3